Use fabsf on floats, fix empty body in if statements.

+++ src/player/view.c
@@ -1164,7 +1164,7 @@ void ClientEndServerFrame (edict_t *ent)
 		bobtime *= 4;
 
 	bobcycle = (int)bobtime;
-	bobfracsin = fabsf(sin(bobtime*M_PI));
+	bobfracsin = fabsf(sinf(bobtime * M_PI));
 
 	// detect hitting the floor
 	P_FallingDamage (ent);
